+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # OCaml Alcotest Test Explorer for Visual Studio Code Changelog
 
+## Version 0.7.0 (2023-03-27)
+
+- If the dune lock can't be acquired for 30s, raise an error window and ask the user what to do.
+- Update the documentation.
+
+### Bugfixes
+
+- Make `Cancel Test Run` stop Dune processes waiting for the lock too. See [Issue #9](https://github.com/Release-Candidate/vscode-ocaml-expect-inline/issues/9).
+
+### Internal Changes
+
+- Add yarn target for the Open VSX Registry.
+
 ## Version 0.6.0 (2023-03-23)
 
 - Make the path to the Dune executable configurable. Can now be either an absolute path, a path relative to the project root or just `dune`, which is looked up in the local Opam environment or the `PATH`.